requirements.txt
openai
langchain
python-dotenv
requests
boto3
fastapi
python-multipart
uvicorn
python-decouple
psycopg2-binary
python-multipart
pyngrok
colorlog
sqlalchemy
sqlalchemy_utils

icalendar

python-jose[cryptography]
six